fix(data): move derive helper tests with macros
