+++ crates/aether-actor-derive/src/lib.rs
@@ -17,77 +17,10 @@ use proc_macro2::TokenStream as TokenStream2;
 use quote::quote;
 use syn::meta;
 use syn::parse::Parser;
-use syn::spanned::Spanned;
 use syn::{
     Attribute, Expr, ExprLit, Fields, FnArg, GenericArgument, ImplItem, ItemImpl, ItemStruct, Lit,
     Meta, PathArguments, ReturnType, Signature, Type, parse_macro_input,
 };
-
-/// Walk a field-type syntactic tree and reject `HashMap` anywhere
-/// inside it (issue #232). `HashMap`'s iteration order is hash-state-
-/// dependent, which would let two builds of the same kind hash to
-/// different `Kind::ID`s — kind ids are derived from canonical schema
-/// bytes, so platform-dependent encoding is a wire-correctness bug.
-/// `BTreeMap` (sorted by key) is the deterministic alternative; the
-/// error message names it explicitly so the fix is one substitution.
-///
-/// Recurses through `AngleBracketed` generic args so nested forms like
-/// `Vec<HashMap<String, String>>` and `Option<HashMap<...>>` are
-/// caught too — the nested case would otherwise pass through trait
-/// dispatch and emit a less actionable "trait `Schema` not
-/// implemented" error pointing at the inner `HashMap`. The walk is
-/// total: it also recurses through array, slice, reference, tuple,
-/// paren, and group positions so a `HashMap` hidden inside `[_; N]`,
-/// `&[_]`, or `(_, _)` earns the same pointed error rather than the
-/// opaque downstream one. A set type (`HashSet`/`BTreeSet`) has no
-/// `Schema` impl at all, so it is rejected here too with a redirect to
-/// a sorted `Vec<T>`.
-fn reject_hashmap(ty: &Type) -> syn::Result<()> {
-    match ty {
-        Type::Path(tp) => {
-            if let Some(seg) = tp.path.segments.last() {
-                if seg.ident == "HashMap" {
-                    return Err(syn::Error::new_spanned(
-                        ty,
-                        "HashMap is not allowed in derived kind schemas — its iteration order is \
-                         platform-dependent and would diverge canonical schema bytes (and Kind::ID) \
-                         across builds. Use `std::collections::BTreeMap` instead, which sorts by key. \
-                         See https://github.com/iamacoffeepot/aether/issues/232",
-                    ));
-                }
-                if seg.ident == "HashSet" || seg.ident == "BTreeSet" {
-                    return Err(syn::Error::new_spanned(
-                        ty,
-                        "a set has no `Schema` impl and is not allowed in derived kind schemas — \
-                         model it as a sorted `Vec<T>`, which encodes deterministically. \
-                         See https://github.com/iamacoffeepot/aether/issues/232",
-                    ));
-                }
-            }
-            for seg in &tp.path.segments {
-                if let PathArguments::AngleBracketed(args) = &seg.arguments {
-                    for arg in &args.args {
-                        if let GenericArgument::Type(inner) = arg {
-                            reject_hashmap(inner)?;
-                        }
-                    }
-                }
-            }
-        }
-        Type::Array(a) => reject_hashmap(&a.elem)?,
-        Type::Slice(s) => reject_hashmap(&s.elem)?,
-        Type::Reference(r) => reject_hashmap(&r.elem)?,
-        Type::Paren(p) => reject_hashmap(&p.elem)?,
-        Type::Group(g) => reject_hashmap(&g.elem)?,
-        Type::Tuple(t) => {
-            for elem in &t.elems {
-                reject_hashmap(elem)?;
-            }
-        }
-        _ => {}
-    }
-    Ok(())
-}
 
 fn to_screaming_snake_case(s: &str) -> String {
     let mut out = String::with_capacity(s.len() + 4);
@@ -3628,123 +3561,5 @@ fn extract_agent_doc(attrs: &[Attribute]) -> Option<String> {
         if s.is_empty() { None } else { Some(s) }
     } else {
         Some(full_trimmed.to_string())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::reject_hashmap;
-    use syn::parse_str;
-
-    // Issue #232: pin the HashMap rejection so a future field-walker
-    // refactor can't silently drop the check (the user-visible
-    // failure mode would be a confusing "Schema not implemented"
-    // error pointing at HashMap rather than the actionable
-    // "use BTreeMap" message). Each fixture covers one shape we
-    // expect the rejection to catch — direct, nested in Vec, nested
-    // in Option, and inside a fully-qualified path.
-
-    fn err(ty: &str) -> String {
-        let parsed: syn::Type = parse_str(ty).expect("test fixture parses");
-        reject_hashmap(&parsed)
-            .err()
-            .unwrap_or_else(|| panic!("expected reject_hashmap to error on {ty}"))
-            .to_string()
-    }
-
-    #[test]
-    fn rejects_direct_hashmap_field() {
-        let msg = err("HashMap<String, String>");
-        assert!(
-            msg.contains("BTreeMap"),
-            "error must point to BTreeMap fix, got: {msg}"
-        );
-        assert!(
-            msg.contains("232"),
-            "error must reference issue 232, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn rejects_fully_qualified_hashmap() {
-        let msg = err("std::collections::HashMap<String, u32>");
-        assert!(msg.contains("BTreeMap"));
-    }
-
-    #[test]
-    fn rejects_hashmap_nested_in_vec() {
-        let msg = err("Vec<HashMap<String, String>>");
-        assert!(msg.contains("BTreeMap"));
-    }
-
-    #[test]
-    fn rejects_hashmap_nested_in_option() {
-        let msg = err("Option<HashMap<String, String>>");
-        assert!(msg.contains("BTreeMap"));
-    }
-
-    #[test]
-    fn rejects_hashmap_in_array() {
-        let msg = err("[HashMap<String, u32>; 4]");
-        assert!(msg.contains("BTreeMap"));
-        assert!(msg.contains("232"));
-    }
-
-    #[test]
-    fn rejects_hashmap_in_tuple() {
-        let msg = err("(u32, HashMap<String, u32>)");
-        assert!(msg.contains("BTreeMap"));
-        assert!(msg.contains("232"));
-    }
-
-    #[test]
-    fn rejects_hashmap_behind_slice_ref() {
-        let msg = err("&[HashMap<String, u32>]");
-        assert!(msg.contains("BTreeMap"));
-        assert!(msg.contains("232"));
-    }
-
-    #[test]
-    fn rejects_hashmap_behind_ref() {
-        let msg = err("&HashMap<String, u32>");
-        assert!(msg.contains("BTreeMap"));
-        assert!(msg.contains("232"));
-    }
-
-    #[test]
-    fn rejects_hashset() {
-        let msg = err("HashSet<u64>");
-        assert!(
-            msg.contains("Vec"),
-            "error must redirect to a sorted Vec, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn rejects_btreeset() {
-        let msg = err("BTreeSet<u64>");
-        assert!(
-            msg.contains("Vec"),
-            "error must redirect to a sorted Vec, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn allows_btreemap_field() {
-        let parsed: syn::Type =
-            parse_str("BTreeMap<String, String>").expect("test setup: BTreeMap type parses");
-        assert!(reject_hashmap(&parsed).is_ok());
-    }
-
-    #[test]
-    fn allows_plain_types() {
-        for ty in ["u32", "String", "Vec<u8>", "Option<String>"] {
-            let parsed: syn::Type =
-                parse_str(ty).expect("test setup: candidate type parses as syn::Type");
-            assert!(
-                reject_hashmap(&parsed).is_ok(),
-                "rejected {ty} unexpectedly"
-            );
-        }
     }
 }
